分类管理三级列表展示、类型管理修改页面传参，默认勾选待开发，GoodsbrandController.checkedById方法

+++ shopping/src/main/webapp/WEB-INF/jsp/goods_brand_list.jsp
@@ -54,6 +54,10 @@
 						str += ",";
 					}
 				}
+				var f = confirm('是否确认删除');
+				if (!f) {
+					return ;
+				}
 				window.location.href="goodsbranddeleteall.htm?ids="+str;
 			}
 			//修改旗
