+++ shopping/src/main/webapp/WEB-INF/jsp/spec_list.jsp
@@ -55,6 +55,10 @@
 						str += ",";
 					}
 				}
+				var f = confirm('是否确认删除');
+				if (!f) {
+					return ;
+				}
 				window.location.href="specdeleteall.htm?ids="+str;
 			}
 		</script>
